Multi vendor entity manager factory

+++ common-spring-jpa-starter/src/main/java/com/olsonsolution/common/spring/application/jpa/config/DataSourceSpecConfig.java
@@ -76,7 +76,7 @@ public class DataSourceSpecConfig {
         int maximumDataSources = routingDataSourceProperties.getMaxDataSources();
         CachingProperties cachingProperties = new DefaultCachingProperties(
                 maximumDataSources,
-                maximumDataSources,
+                (long) maximumDataSources,
                 null,
                 routingDataSourceProperties.getExpireTimeout(),
                 null
@@ -96,7 +96,7 @@ public class DataSourceSpecConfig {
         int maximumDataSources = routingDataSourceProperties.getMaxDataSources();
         CachingProperties cachingProperties = new DefaultCachingProperties(
                 maximumDataSources,
-                maximumDataSources,
+                (long) maximumDataSources,
                 null,
                 routingDataSourceProperties.getExpireTimeout(),
                 null
